# swagger.json
{
    "swagger": "2.0",
    "info": {
    "version": "1.0.0",
    "title": "REST API proyectoUTM",
    "description": "<html><body>Esta es la documentación de la API para el sistema de Investigación Docencia.<p>Para poder usar esta API es necesario solicitar el token y con esto poder utilizar los servicios proporcionados.</p><br><p><b>Autenticación y token de API</b></p><p>La Universidad Tecnológica de la Mixteca (UTM) proporciona el api-token. Una vez que haya creado y confirmado su cuenta.</p></body></html>",
    "license": {
    "name": "MIT",
    "url": "https://opensource.org/licenses/MIT"
    }
    },
    "host": "localhost:3100",
    "basePath": "",
...
    "schemes": [
        "http"
        ],
        "consumes":[
        "application/json"
        ],
        "produces" :[
        "application/json"
        ],
        "paths":
...
            "/api/institutos/actualizar/{idInstituto}":
            {
                "put":
                {
                    "summary": "Modificando un instituto",
                    "operationId": "actualizar",
                    "consumes": [ "application/json" ],
                    "produces": [
                        "application/json"
                    ],
                    "parameters": [
                        {
                            "name": "idInstituto",
                            "in": "path",
                            "description": "id del Instituto a eliminar",
                            "required": true,
                            "default":18
                        },
                        {
                            "name": "Instituto",
                            "in": "body",
                            "description": "Instituto a modificar",
                            "required": true,
                            "schema": {
                                "$ref": "#/definitions/ViejoInstituto"
                            }
                        }
                    ],
                    "tags": [ "Institutos" ],
                    "responses": {
                        "201": {
                            "description": "respuesta vacía"
                        },
                        "403": {
                            "description": "Prohibido"
                        }
                    }
                }
            },
...
            "/api/carreras/actualizar/{idCarrera}":
            {
                "put":
                {
                    "summary": "Modificando una carrera",
                    "operationId": "actualizar",
                    "consumes": [ "application/json" ],
                    "produces": [
                        "application/json"
                    ],
                    "parameters": [
                        {
                            "name": "idCarrera",
                            "in": "path",
                            "description": "id de la carrera a eliminar",
                            "required": true,
                            "default":18
                        },
                        {
                            "name": "Carrera",
                            "in": "body",
                            "description": "Carrera a modificar",
                            "required": true,
                            "schema": {
                                "$ref": "#/definitions/ViejaCarrera"
                            }
                        }
                    ],
                    "tags": [ "Carreras" ],
                    "responses": {
                        "201": {
                            "description": "respuesta vacía"
                        },
                        "403": {
                            "description": "Prohibido"
                        }
                    }
                }
            },
...
            "/api/tipoProfesor/actualizar/{idTipoProfesor}":
            {
                "put":
                {
                    "summary": "Modificando un tipo de profesor",
                    "operationId": "actualizar",
                    "consumes": [ "application/json" ],
                    "produces": [
                        "application/json"
                    ],
                    "parameters": [
                        {
                            "name": "idTipoProfesor",
                            "in": "path",
                            "description": "id del tipo de profesor a eliminar",
                            "required": true,
                            "default":18
                        },
                        {
                            "name": "tipoProfesor",
                            "in": "body",
                            "description": "tipo Profesor a modificar",
                            "required": true,
                            "schema": {
                                "$ref": "#/definitions/ViejoTipoProfesor"
                            }
                        }
                    ],
                    "tags": [ "Tipo Profesor" ],
                    "responses": {
                        "201": {
                            "description": "respuesta vacía"
                        },
                        "403": {
                            "description": "Prohibido"
                        }
                    }
                }
            },
...
            "/api/profesores/actualizar/{idProfesor}":
            {
                "put":
                {
                    "summary": "Modificando un profesor",
                    "operationId": "actualizar",
                    "consumes": [ "application/json" ],
                    "produces": [
                        "application/json"
                    ],
                    "parameters": [
                        {
                            "name": "idProfesor",
                            "in": "path",
                            "description": "id del profesor a eliminar",
                            "required": true,
                            "default":18
                        },
                        {
                            "name": "Profesor",
                            "in": "body",
                            "description": "Profesor a modificar",
                            "required": true,
                            "schema": {
                                "$ref": "#/definitions/ViejoProfesor"
                            }
                        }
                    ],
                    "tags": [ "Profesores" ],
                    "responses": {
                        "201": {
                            "description": "respuesta vacía"
                        },
                        "403": {
                            "description": "Prohibido"
                        }
                    }
                }
            },
...
        },
         "definitions":
...
            "ViejoInstituto":
            {
                "type": "object",
                "required": [
                    "codigoInstituto","nombre"
                ],
                "properties": {
                    "codigoInstituto":
                    {
                        "type": "string",
                        "default":"111"
                    },
                    "nombre": {
                        "type": "string",
                        "default":"Instituto patito"
                    }
                }
            },
...
            "ViejaCarrera":
            {
                "type": "object",
                "required": [
                    "codigoCarrera","nombreCarrera","idInstituto"
                ],
                "properties": {
                    "codigoCarrera":
                    {
                        "type": "string",
                        "default":"111"
                    },
                    "nombreCarrera": {
                        "type": "string",
                        "default":"Carrera patito"
                    },
                    "idInstituto": {
                        "type": "number",
                        "default":"2"
                    }
                }
            },
...
            "ViejoTipoProfesor":
            {
                "type": "object",
                "required": [
                    "idTipoProfesor","nombreTipo"
                ],
                "properties": {
                    "idTipoProfesor":
                    {
                        "type": "number",
                        "default":"111"
                    },
                    "nombreTipo": {
                        "type": "string",
                        "default":"Tipo patito"
                    }
                }
            },
...
            "ViejoProfesor":
            {
                "type": "object",
                "required": [
                    "idProfesor","nombres","apellidoP","apellidoM","correoProfesor", "password","nivel","idCarrera","grado","idTipoProfesor"
                ],
                "properties": {
                    "idProfesor":
                     {
                        "type": "number",
                        "default":"111"
                    },
                    "nombres":
                    {
                        "type":"string",
                        "default":"Juan "
                    },
                    "apellidoP":
                    {
                        "type":"string",
                        "default":"Cruz"
                    },
                    "apellidoM":
                    {
                        "type":"string",
                        "default":"Perez"
                    },
                    "correoProfesor":
                    {
                        "type":"string",
                        "default":"algo@algomas"
                    },
                    "password":
                    {
                        "type":"string",
                        "default":"algo"
                    },
                    "nivel":
                    {
                        "type":"number",
                        "default":"2"
                    },
                    "idCarrera":
                    {
                        "type":"number",
                        "default":"111"
                    },
                    "grado":
                    {
                        "type":"string",
                        "default":"eso"
                    },
                    "idTipoProfesor":
                    {
                        "type":"number",
                        "default":"3"
                    }
                }
            },
...
        }
}